+++ src/components/LocationDropDown.js
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import { useContext } from 'react';
 import { LocationContext } from '../contexts/LocationContext';
 
 function DropDown() {
